+++ bittytax/conv/parsers/__init__.py
@@ -23,6 +23,7 @@ from . import hitbtc
 from . import hotbit
 from . import ii
 from . import kucoin
+from . import kraken
 from . import ledgerlive
 from . import liquid
 from . import nexo