+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Added
 - Conversion tool: added parser for CGTCalculator.
 - Conversion tool: added parser for Nexo.
+- Conversion tool: added parser for Kraken.
 ### Changed
 - Hotbit parser: Negative fees are now set to zero.
 - Accounting tool: Drop buy/sell/fee transactions of zero quantity.